% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_package.R
\docType{package}
\name{crew.aws.batch-package}
\alias{crew.aws.batch-package}
\title{crew.aws.batch: a crew launcher plugin for AWS Batch}
\description{
In computationally demanding analysis projects,
statisticians and data scientists asynchronously
deploy long-running tasks to distributed systems,
ranging from traditional clusters to cloud services.
The \code{crew.aws.batch} package extends the
\href{https://github.com/shikokuchuo/mirai}{\code{mirai}}-powered
\href{https://wlandau.github.io}{\code{crew}} package with worker
launcher plugins for AWS Batch.
Inspiration also comes from packages
\href{https://github.com/shikokuchuo/mirai}{\code{mirai}},
\href{https://future.futureverse.org/}{\code{future}},
\href{https://mrc-ide.github.io/rrq/}{\code{rrq}},
\href{https://mschubert.github.io/clustermq/}{\code{clustermq}},
and \href{https://mllg.github.io/batchtools/}{\code{batchtools}}.
}
\concept{help}
