% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_aws_batch_job_definition.R
\name{crew_aws_batch_job_definition}
\alias{crew_aws_batch_job_definition}
\title{Create an AWS Batch job definition object.}
\usage{
crew_aws_batch_job_definition(name = NULL)
}
\arguments{
\item{name}{Character of length 1, job definition name.}
}
\description{
Create an \code{R6} AWS Batch job definition object.
}
\concept{job_definition}
