% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_launcher_aws_batch.R
\name{crew_aws_batch_job_name}
\alias{crew_aws_batch_job_name}
\title{Terminate an AWS Batch job.}
\usage{
crew_aws_batch_job_name(name)
}
\arguments{
\item{name}{Character string, an AWS Batch job name, possibly invalid.}
}
\value{
Character string, a valid AWS Batch job name.
}
\description{
Not a user-side function. For internal use only.
}
\keyword{internal}
