% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aws_batch.R
\name{crew_aws_batch_launch}
\alias{crew_aws_batch_launch}
\title{Submit an AWS Batch job.}
\usage{
crew_aws_batch_launch(args_client, args_submit)
}
\arguments{
\item{args_client}{Named list of arguments to \code{paws.compute::batch()}.}

\item{args_submit}{Named list of arguments to
\code{paws.compute::batch()$submit_job()}.}
}
\value{
HTTP response from submitting the job.
}
\description{
Not a user-side function. For internal use only.
}
\keyword{internal}
