% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_aws_batch_monitor.R
\name{crew_aws_batch_monitor}
\alias{crew_aws_batch_monitor}
\title{Create an AWS Batch monitor object.}
\usage{
crew_aws_batch_monitor(
  job_queue,
  job_definition = paste0("crew-aws-batch-job-definition-", crew::crew_random_name()),
  log_group = "/aws/batch/job",
  log_group_region = "us-east-1"
)
}
\arguments{
\item{job_queue}{Character of length 1, name of the AWS Batch
job queue.}

\item{job_definition}{Character of length 1, name of the AWS Batch
job definition.}

\item{log_group}{Character of length 1,
AWS Batch CloudWatch log group to get job logs.
The default log group is often "/aws/batch/job", but not always.
It is not easy to get the log group of an active job or job
definition, so if you have a non-default log group and you do not
know its name, please consult your system administrator.}

\item{log_group_region}{Character of length 1, region of the log group.
Must be a valid region name such as \code{"us-east-1"} or \code{"us-east-2"}.}
}
\description{
Create an \code{R6} object to manage AWS Batch jobs and
job definitions.
}
\seealso{
Other monitor: 
\code{\link{crew_class_aws_batch_monitor}}
}
\concept{monitor}
