% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_aws_batch_monitor.R
\name{crew_class_aws_batch_monitor}
\alias{crew_class_aws_batch_monitor}
\title{AWS Batch monitor class}
\description{
AWS Batch job definition \code{R6} class
}
\details{
See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.
}
\seealso{
Other monitor: 
\code{\link{crew_aws_batch_monitor}()}
}
\concept{monitor}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{job_queue}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}

\item{\code{job_definition}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}

\item{\code{log_group}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}

\item{\code{config}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}

\item{\code{credentials}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}

\item{\code{endpoint}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}

\item{\code{region}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_aws_batch_monitor-new}{\code{crew_class_aws_batch_monitor$new()}}
\item \href{#method-crew_class_aws_batch_monitor-validate}{\code{crew_class_aws_batch_monitor$validate()}}
\item \href{#method-crew_class_aws_batch_monitor-register}{\code{crew_class_aws_batch_monitor$register()}}
\item \href{#method-crew_class_aws_batch_monitor-deregister}{\code{crew_class_aws_batch_monitor$deregister()}}
\item \href{#method-crew_class_aws_batch_monitor-describe}{\code{crew_class_aws_batch_monitor$describe()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_aws_batch_monitor-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_aws_batch_monitor-new}{}}}
\subsection{Method \code{new()}}{
AWS Batch job definition constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_aws_batch_monitor$new(
  job_queue = NULL,
  job_definition = NULL,
  log_group = NULL,
  config = NULL,
  credentials = NULL,
  endpoint = NULL,
  region = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{job_queue}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}

\item{\code{job_definition}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}

\item{\code{log_group}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}

\item{\code{config}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}

\item{\code{credentials}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}

\item{\code{endpoint}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}

\item{\code{region}}{See \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AWS Batch job definition object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_aws_batch_monitor-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_aws_batch_monitor-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_aws_batch_monitor$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly). Throws an error if a field is invalid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_aws_batch_monitor-register"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_aws_batch_monitor-register}{}}}
\subsection{Method \code{register()}}{
Register a job definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_aws_batch_monitor$register(
  image,
  platform_capabilities = "EC2",
  memory_units = "gigabytes",
  memory = NULL,
  cpus = NULL,
  gpus = NULL,
  seconds_timeout = NULL,
  scheduling_priority = NULL,
  tags = NULL,
  propagate_tags = NULL,
  job_role_arn = NULL,
  execution_role_arn = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image}}{Character of length 1, Docker image used for each job.
You can supply a path to an image in Docker Hub or the full URI
of an image in an Amazohn ECR repository.}

\item{\code{platform_capabilities}}{Optional character of length 1, either
\code{"EC2"} to run on EC2 or \code{"FARGATE"} to run on Fargate.}

\item{\code{memory_units}}{Character of length 1,
either \code{"gigabytes"} or \code{"mebibytes"} to set the units of the
\code{memory} argument. \code{"gigabytes"} is simpler for EC2 jobs, but
Fargate has strict requirements about specifying exact amounts of
mebibytes (MiB). for details, read
\url{https://docs.aws.amazon.com/cli/latest/reference/batch/register-job-definition.html} # nolint}

\item{\code{memory}}{Positive numeric of length 1, amount of memory to request
for each job.}

\item{\code{cpus}}{Positive numeric of length 1, number of virtual
CPUs to request for each job.}

\item{\code{gpus}}{Positive numeric of length 1, number of GPUs to
request for each job.}

\item{\code{seconds_timeout}}{Optional positive numeric of length 1,
number of seconds until a job times out.}

\item{\code{scheduling_priority}}{Optional nonnegative integer of length 1
between \code{0} and \code{9999}, priority of jobs. Higher-valued jobs
are scheduled first. Only applies if the job queue has a fair share
policy. Set to \code{NULL} to omit.
which has a fair share policy.}

\item{\code{tags}}{Optional character vector of tags.}

\item{\code{propagate_tags}}{Optional logical of length 1, whether to propagate
tags from the job or definition to the ECS task.}

\item{\code{job_role_arn}}{Character of length 1,
Amazon resource name (ARN) of the job role.}

\item{\code{execution_role_arn}}{Character of length 1,
Amazon resource name (ARN) of the execution role.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The \code{register()} method registers a simple
job definition using the job definition name and log group originally
supplied to \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.
Job definitions created with \verb{$register()} are container-based
and use the AWS log driver.
For more complicated
kinds of jobs, we recommend skipping \code{register()}: first call
\url{https://www.paws-r-sdk.com/docs/batch_register_job_definition/}
to register the job definition, then supply the job definition
name to the \code{job_definition} argument of \code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.
}

\subsection{Returns}{
A one-row \code{tibble} with the job definition name, ARN, and
revision number of the registered job definition.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_aws_batch_monitor-deregister"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_aws_batch_monitor-deregister}{}}}
\subsection{Method \code{deregister()}}{
Attempt to deeregister the job definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_aws_batch_monitor$deregister()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Attempt to dereregister the job definition whose name was
originally supplied to the \code{job_definition} argument of
\code{\link[=crew_aws_batch_monitor]{crew_aws_batch_monitor()}}.
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_aws_batch_monitor-describe"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_aws_batch_monitor-describe}{}}}
\subsection{Method \code{describe()}}{
Describe the current active revision of the job definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_aws_batch_monitor$describe()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
If the job definition is not active or does not exist,
\code{describe()} returns \code{NULL}. Otherwise, it returns
a \code{tibble} with job definition information. Some fields
may be nested lists.
}
}
}
