% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_monitor_aws_batch.R
\name{crew_class_monitor_aws_batch}
\alias{crew_class_monitor_aws_batch}
\title{AWS Batch monitor class}
\description{
AWS Batch job definition \code{R6} class
}
\details{
See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
}
\seealso{
Other monitor: 
\code{\link{crew_monitor_aws_batch}()}
}
\concept{monitor}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{job_queue}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}

\item{\code{job_definition}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}

\item{\code{log_group}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}

\item{\code{config}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}

\item{\code{credentials}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}

\item{\code{endpoint}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}

\item{\code{region}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_monitor_aws_batch-new}{\code{crew_class_monitor_aws_batch$new()}}
\item \href{#method-crew_class_monitor_aws_batch-validate}{\code{crew_class_monitor_aws_batch$validate()}}
\item \href{#method-crew_class_monitor_aws_batch-register}{\code{crew_class_monitor_aws_batch$register()}}
\item \href{#method-crew_class_monitor_aws_batch-deregister}{\code{crew_class_monitor_aws_batch$deregister()}}
\item \href{#method-crew_class_monitor_aws_batch-describe}{\code{crew_class_monitor_aws_batch$describe()}}
\item \href{#method-crew_class_monitor_aws_batch-submit}{\code{crew_class_monitor_aws_batch$submit()}}
\item \href{#method-crew_class_monitor_aws_batch-terminate}{\code{crew_class_monitor_aws_batch$terminate()}}
\item \href{#method-crew_class_monitor_aws_batch-status}{\code{crew_class_monitor_aws_batch$status()}}
\item \href{#method-crew_class_monitor_aws_batch-log}{\code{crew_class_monitor_aws_batch$log()}}
\item \href{#method-crew_class_monitor_aws_batch-jobs}{\code{crew_class_monitor_aws_batch$jobs()}}
\item \href{#method-crew_class_monitor_aws_batch-active}{\code{crew_class_monitor_aws_batch$active()}}
\item \href{#method-crew_class_monitor_aws_batch-inactive}{\code{crew_class_monitor_aws_batch$inactive()}}
\item \href{#method-crew_class_monitor_aws_batch-submitted}{\code{crew_class_monitor_aws_batch$submitted()}}
\item \href{#method-crew_class_monitor_aws_batch-pending}{\code{crew_class_monitor_aws_batch$pending()}}
\item \href{#method-crew_class_monitor_aws_batch-runnable}{\code{crew_class_monitor_aws_batch$runnable()}}
\item \href{#method-crew_class_monitor_aws_batch-starting}{\code{crew_class_monitor_aws_batch$starting()}}
\item \href{#method-crew_class_monitor_aws_batch-running}{\code{crew_class_monitor_aws_batch$running()}}
\item \href{#method-crew_class_monitor_aws_batch-succeeded}{\code{crew_class_monitor_aws_batch$succeeded()}}
\item \href{#method-crew_class_monitor_aws_batch-failed}{\code{crew_class_monitor_aws_batch$failed()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-new}{}}}
\subsection{Method \code{new()}}{
AWS Batch job definition constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$new(
  job_queue = NULL,
  job_definition = NULL,
  log_group = NULL,
  config = NULL,
  credentials = NULL,
  endpoint = NULL,
  region = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{job_queue}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}

\item{\code{job_definition}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}

\item{\code{log_group}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}

\item{\code{config}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}

\item{\code{credentials}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}

\item{\code{endpoint}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}

\item{\code{region}}{See \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AWS Batch job definition object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly). Throws an error if a field is invalid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-register"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-register}{}}}
\subsection{Method \code{register()}}{
Register a job definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$register(
  image,
  platform_capabilities = "EC2",
  memory_units = "gigabytes",
  memory = NULL,
  cpus = NULL,
  gpus = NULL,
  seconds_timeout = NULL,
  scheduling_priority = NULL,
  tags = NULL,
  propagate_tags = NULL,
  parameters = NULL,
  job_role_arn = NULL,
  execution_role_arn = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{image}}{Character of length 1, Docker image used for each job.
You can supply a path to an image in Docker Hub or the full URI
of an image in an Amazon ECR repository.}

\item{\code{platform_capabilities}}{Optional character of length 1, either
\code{"EC2"} to run on EC2 or \code{"FARGATE"} to run on Fargate.}

\item{\code{memory_units}}{Character of length 1,
either \code{"gigabytes"} or \code{"mebibytes"} to set the units of the
\code{memory} argument. \code{"gigabytes"} is simpler for EC2 jobs, but
Fargate has strict requirements about specifying exact amounts of
mebibytes (MiB). for details, read
\url{https://docs.aws.amazon.com/cli/latest/reference/batch/register-job-definition.html} # nolint}

\item{\code{memory}}{Positive numeric of length 1, amount of memory to request
for each job.}

\item{\code{cpus}}{Positive numeric of length 1, number of virtual
CPUs to request for each job.}

\item{\code{gpus}}{Positive numeric of length 1, number of GPUs to
request for each job.}

\item{\code{seconds_timeout}}{Optional positive numeric of length 1,
number of seconds until a job times out.}

\item{\code{scheduling_priority}}{Optional nonnegative integer of length 1
between \code{0} and \code{9999}, priority of jobs. Jobs with higher-valued
priorities are scheduled first.
The priority only applies if the job queue has a fair share
policy. Set to \code{NULL} to omit.}

\item{\code{tags}}{Optional character vector of tags.}

\item{\code{propagate_tags}}{Optional logical of length 1, whether to propagate
tags from the job or definition to the ECS task.}

\item{\code{parameters}}{Optional character vector of key-value pairs
designating parameters for job submission.}

\item{\code{job_role_arn}}{Character of length 1,
Amazon resource name (ARN) of the job role.}

\item{\code{execution_role_arn}}{Character of length 1,
Amazon resource name (ARN) of the execution role.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The \code{register()} method registers a simple
job definition using the job definition name and log group originally
supplied to \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
Job definitions created with \verb{$register()} are container-based
and use the AWS log driver.
For more complicated
kinds of jobs, we recommend skipping \code{register()}: first call
\url{https://www.paws-r-sdk.com/docs/batch_register_job_definition/}
to register the job definition, then supply the job definition
name to the \code{job_definition} argument of \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
}

\subsection{Returns}{
A one-row \code{tibble} with the job definition name, ARN, and
revision number of the registered job definition.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-deregister"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-deregister}{}}}
\subsection{Method \code{deregister()}}{
Attempt to deregister the job definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$deregister()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Attempt to deregister the job definition whose name was
originally supplied to the \code{job_definition} argument of
\code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
}

\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-describe"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-describe}{}}}
\subsection{Method \code{describe()}}{
Describe the current active revision of the job definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$describe()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
If the job definition is not active or does not exist,
\code{describe()} returns \code{NULL}. Otherwise, it returns
a \code{tibble} with job definition information. Some fields
may be nested lists.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-submit"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-submit}{}}}
\subsection{Method \code{submit()}}{
Submit a single AWS Batch job to the given job queue
under the given job definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$submit(
  command = c("sleep", "300"),
  name = paste0("crew-aws-batch-job-", crew::crew_random_name()),
  memory_units = "gigabytes",
  memory = NULL,
  cpus = NULL,
  gpus = NULL,
  seconds_timeout = NULL,
  share_identifier = NULL,
  scheduling_priority_override = NULL,
  tags = NULL,
  propagate_tags = NULL,
  parameters = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{command}}{Character vector with the command
to submit for the job. Usually a Linux shell command
with each term in its own character string.}

\item{\code{name}}{Character of length 1 with the job name.}

\item{\code{memory_units}}{Character of length 1,
either \code{"gigabytes"} or \code{"mebibytes"} to set the units of the
\code{memory} argument. \code{"gigabytes"} is simpler for EC2 jobs, but
Fargate has strict requirements about specifying exact amounts of
mebibytes (MiB). for details, read
\url{https://docs.aws.amazon.com/cli/latest/reference/batch/register-job-definition.html} # nolint}

\item{\code{memory}}{Positive numeric of length 1, amount of memory to request
for each job.}

\item{\code{cpus}}{Positive numeric of length 1, number of virtual
CPUs to request for each job.}

\item{\code{gpus}}{Positive numeric of length 1, number of GPUs to
request for each job.}

\item{\code{seconds_timeout}}{Optional positive numeric of length 1,
number of seconds until a job times out.}

\item{\code{share_identifier}}{Character of length 1 with the share
identifier of the job. Only applies if the job queue has a
scheduling policy. Read the official AWS Batch documentation
for details.}

\item{\code{scheduling_priority_override}}{Optional nonnegative integer
of length between \code{0} and \code{9999}, priority of the job.
This value overrides the priority in the job definition.
Jobs with higher-valued priorities are scheduled first.
The priority applies if the job queue has a fair share policy.
Set to \code{NULL} to omit.}

\item{\code{tags}}{Optional character vector of tags.}

\item{\code{propagate_tags}}{Optional logical of length 1, whether to propagate
tags from the job or definition to the ECS task.}

\item{\code{parameters}}{Optional character vector of key-value pairs
designating parameters for job submission.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method uses the job queue and job definition
that were supplied through \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
Any jobs submitted this way are different from the
\code{crew} workers that the \code{crew} controller starts automatically
using the AWS Batch launcher plugin.
You may use the \code{submit()} method in the monitor for different
purposes such as testing.
}

\subsection{Returns}{
A one-row \code{tibble} with the name, ID, and
Amazon resource name (ARN) of the job.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-terminate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-terminate}{}}}
\subsection{Method \code{terminate()}}{
Terminate one or more AWS Batch jobs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$terminate(
  ids,
  reason = "terminated by crew.aws.batch monitor",
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{Character vector with the IDs of the AWS Batch jobs
to terminate.}

\item{\code{reason}}{Character of length 1, natural language explaining
the reason the job was terminated.}

\item{\code{verbose}}{Logical of length 1, whether to show a progress bar
if the R process is interactive and \code{length(ids)} is greater than 1.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-status"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-status}{}}}
\subsection{Method \code{status()}}{
Get the status of a single job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$status(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Character of length 1, job ID. This is different
from the user-supplied job name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A one-row \code{tibble} with information about the job.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-log"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-log}{}}}
\subsection{Method \code{log()}}{
Get the CloudWatch log of a job.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$log(id, start_from_head = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Character of length 1, job ID. This is different
from the user-supplied job name.}

\item{\code{start_from_head}}{Logical of length 1, whether to print earlier
log events before later ones.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method assumes the job has log driver \code{"awslogs"}
(specifying AWS CloudWatch) and that the log group is the one
prespecified in the \code{log_group} argument of
\code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}. This method cannot use
other log drivers such as Splunk, and it will fail if the log
group is wrong or missing.
}

\subsection{Returns}{
A \code{tibble} with log information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-jobs"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-jobs}{}}}
\subsection{Method \code{jobs()}}{
List all the jobs in the given job queue
with the given job definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$jobs(
  status = c("submitted", "pending", "runnable", "starting", "running", "succeeded",
    "failed")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{status}}{Character vector of job states. Results are limited
to these job states.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The output only includes jobs under the
job queue and job definition
that were supplied through \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
}

\subsection{Returns}{
A \code{tibble} with one row per job and columns
with job information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-active"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-active}{}}}
\subsection{Method \code{active()}}{
List active jobs: submitted, pending,
runnable, starting, or running (not succeeded or failed).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$active()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The output only includes jobs under the
job queue and job definition
that were supplied through \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
}

\subsection{Returns}{
A \code{tibble} with one row per job and columns
with job information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-inactive"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-inactive}{}}}
\subsection{Method \code{inactive()}}{
List inactive jobs: ones whose status
is succeeded or failed (not submitted, pending,
runnable, starting, or running).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$inactive()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The output only includes jobs under the
job queue and job definition
that were supplied through \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
}

\subsection{Returns}{
A \code{tibble} with one row per job and columns
with job information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-submitted"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-submitted}{}}}
\subsection{Method \code{submitted()}}{
List jobs whose status is \code{"submitted"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$submitted()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The output only includes jobs under the
job queue and job definition
that were supplied through \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
}

\subsection{Returns}{
A \code{tibble} with one row per job and columns
with job information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-pending"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-pending}{}}}
\subsection{Method \code{pending()}}{
List jobs whose status is \code{"pending"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$pending()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The output only includes jobs under the
job queue and job definition
that were supplied through \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
}

\subsection{Returns}{
A \code{tibble} with one row per job and columns
with job information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-runnable"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-runnable}{}}}
\subsection{Method \code{runnable()}}{
List jobs whose status is \code{"runnable"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$runnable()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The output only includes jobs under the
job queue and job definition
that were supplied through \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
}

\subsection{Returns}{
A \code{tibble} with one row per job and columns
with job information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-starting"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-starting}{}}}
\subsection{Method \code{starting()}}{
List jobs whose status is \code{"starting"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$starting()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The output only includes jobs under the
job queue and job definition
that were supplied through \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
}

\subsection{Returns}{
A \code{tibble} with one row per job and columns
with job information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-running"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-running}{}}}
\subsection{Method \code{running()}}{
List jobs whose status is \code{"running"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$running()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The output only includes jobs under the
job queue and job definition
that were supplied through \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
}

\subsection{Returns}{
A \code{tibble} with one row per job and columns
with job information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-succeeded"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-succeeded}{}}}
\subsection{Method \code{succeeded()}}{
List jobs whose status is \code{"succeeded"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$succeeded()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The output only includes jobs under the
job queue and job definition
that were supplied through \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
}

\subsection{Returns}{
A \code{tibble} with one row per job and columns
with job information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_monitor_aws_batch-failed"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_monitor_aws_batch-failed}{}}}
\subsection{Method \code{failed()}}{
List jobs whose status is \code{"failed"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_monitor_aws_batch$failed()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The output only includes jobs under the
job queue and job definition
that were supplied through \code{\link[=crew_monitor_aws_batch]{crew_monitor_aws_batch()}}.
}

\subsection{Returns}{
A \code{tibble} with one row per job and columns
with job information.
}
}
}
