% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_launcher_aws_batch.R
\name{crew_launcher_aws_batch_launch}
\alias{crew_launcher_aws_batch_launch}
\title{Submit an AWS Batch job.}
\usage{
crew_launcher_aws_batch_launch(args_client, args_submit)
}
\arguments{
\item{args_client}{Named list of arguments to \code{paws.compute::batch()}.}

\item{args_submit}{Named list of arguments to
\code{paws.compute::batch()$submit_job()}.}
}
\value{
HTTP response from submitting the job.
}
\description{
Not a user-side function. For internal use only.
}
\details{
This utility is its own separate exported function specific to
the launcher and not shared with the job definition or monitor classes.
It generates the \code{paws.compute::batch()} client within itself
instead of a method inside the class.
This is all because it needs to run on a separate local worker process
and it needs to accept exportable arguments.
}
\keyword{internal}
