% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_options_aws_batch.R
\name{crew_options_aws_batch}
\alias{crew_options_aws_batch}
\title{AWS Batch options}
\usage{
crew_options_aws_batch(
  job_definition,
  job_queue,
  cpus = NULL,
  gpus = NULL,
  memory = NULL,
  memory_units = "gigabytes",
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL,
  share_identifier = NULL,
  scheduling_priority_override = NULL,
  parameters = NULL,
  container_overrides = NULL,
  node_overrides = NULL,
  retry_strategy = NULL,
  propagate_tags = NULL,
  timeout = NULL,
  tags = NULL,
  eks_properties_override = NULL
)
}
\arguments{
\item{job_definition}{Character of length 1, name of the AWS
Batch job definition to use. There is no default for this argument,
and a job definition must be created prior to running the controller.
Please see \url{https://docs.aws.amazon.com/batch/} for details.

To create a job definition, you will need to create a Docker-compatible
image which can run R and \code{crew}. You may which to inherit
from the images at \url{https://github.com/rocker-org/rocker-versioned2}.}

\item{job_queue}{Character of length 1, name of the AWS
Batch job queue to use. There is no default for this argument,
and a job queue must be created prior to running the controller.
Please see \url{https://docs.aws.amazon.com/batch/} for details.}

\item{cpus}{Number of virtual CPUs to request per job. Can be \code{NULL}
to go with the defaults in the job definition. Ignored if
\code{container_overrides} is not \code{NULL}.}

\item{gpus}{Number of GPUs to request per job. Can be \code{NULL}
to go with the defaults in the job definition. Ignored if
\code{container_overrides} is not \code{NULL}.}

\item{memory}{Positive number, amount of memory to request per job.
Choose the units of memory with the \code{memory_units} argument.
Fargate instances can only be certain discrete values of mebibytes,
so please choose \code{memory_units = "mebibytes"} in that case.
The \code{memory} argument can be \code{NULL}
to go with the defaults in the job definition. Ignored if
\code{container_overrides} is not \code{NULL}.}

\item{share_identifier}{\code{NULL} or character of length 1.
For details, visit
\url{https://www.paws-r-sdk.com/docs/batch_submit_job/} and the
"AWS arguments" sections of this help file.}

\item{scheduling_priority_override}{\code{NULL} or integer of length 1.
For details, visit
\url{https://www.paws-r-sdk.com/docs/batch_submit_job/} and the
"AWS arguments" sections of this help file.}

\item{parameters}{\code{NULL} or a nonempty list.
For details, visit
\url{https://www.paws-r-sdk.com/docs/batch_submit_job/} and the
"AWS arguments" sections of this help file.}

\item{container_overrides}{\code{NULL} or a nonempty named list of
fields to override
in the container specified in the job definition. Any overrides for the
\code{command} field are ignored because \code{crew.aws.batch} needs to override
the command to run the \code{crew} worker.
For more details, visit
\url{https://www.paws-r-sdk.com/docs/batch_submit_job/} and the
"AWS arguments" sections of this help file.}

\item{node_overrides}{\code{NULL} or a nonempty named list.
For more details, visit
\url{https://www.paws-r-sdk.com/docs/batch_submit_job/} and the
"AWS arguments" sections of this help file.}

\item{retry_strategy}{\code{NULL} or a nonempty named list.
For more details, visit
\url{https://www.paws-r-sdk.com/docs/batch_submit_job/} and the
"AWS arguments" sections of this help file.}

\item{propagate_tags}{\code{NULL} or a logical of length 1.
For more details, visit
\url{https://www.paws-r-sdk.com/docs/batch_submit_job/} and the
"AWS arguments" sections of this help file.}

\item{timeout}{\code{NULL} or a nonempty named list.
For more details, visit
\url{https://www.paws-r-sdk.com/docs/batch_submit_job/} and the
"AWS arguments" sections of this help file.}

\item{tags}{\code{NULL} or a nonempty named list.
For more details, visit
\url{https://www.paws-r-sdk.com/docs/batch_submit_job/} and the
"AWS arguments" sections of this help file.}

\item{eks_properties_override}{\code{NULL} or a nonempty named list.
For more details, visit
\url{https://www.paws-r-sdk.com/docs/batch_submit_job/} and the
"AWS arguments" sections of this help file.}

\item{units_memory}{Character string, units of memory of the \code{memory}
argument. Can be \code{"gigabytes"} or \code{"mebibytes"}.
Fargate instances can only be certain discrete values of mebibytes,
so please choose \code{memory_units = "mebibytes"} in that case.}
}
\value{
A classed list of options for the controller.
}
\description{
Options for the AWS Batch controller.
}
\keyword{plugin_aws_batch}
