% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_aws_batch_job_definition.R
\name{crew_class_aws_batch_job_definition}
\alias{crew_class_aws_batch_job_definition}
\title{AWS Batch job definition class}
\description{
AWS Batch job definition \code{R6} class
}
\details{
See \code{\link[=crew_aws_batch_job_definition]{crew_aws_batch_job_definition()}}.
}
\seealso{
Other plugin_aws_batch: 
\code{\link{crew_class_launcher_aws_batch}},
\code{\link{crew_controller_aws_batch}()},
\code{\link{crew_launcher_aws_batch}()}
}
\concept{plugin_aws_batch}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{See \code{\link[=crew_aws_batch_job_definition]{crew_aws_batch_job_definition()}}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-crew_class_aws_batch_job_definition-new}{\code{crew_class_aws_batch_job_definition$new()}}
\item \href{#method-crew_class_aws_batch_job_definition-validate}{\code{crew_class_aws_batch_job_definition$validate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_aws_batch_job_definition-new"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_aws_batch_job_definition-new}{}}}
\subsection{Method \code{new()}}{
AWS Batch job definition constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_aws_batch_job_definition$new(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{See \code{\link[=crew_aws_batch_job_definition]{crew_aws_batch_job_definition()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AWS Batch job definition object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-crew_class_aws_batch_job_definition-validate"></a>}}
\if{latex}{\out{\hypertarget{method-crew_class_aws_batch_job_definition-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{crew_class_aws_batch_job_definition$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} (invisibly). Throws an error if a field is invalid.
}
}
}
